layout improvements. standardization of styles. finalizing editing and deleting records. inclusion of confirmation dialog.

+++ XFSample/XFSample/ViewModels/PersonDetailViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Input;
@@ -19,31 +18,31 @@ namespace XFSample.ViewModels
 
         public PersonDetailViewModel(INavigation navigation, int id = 0) : base(navigation)
         {
-            _repository =
-                new Repository<Person>
-                (Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "People.db3"));
-
+            _repository = new Repository<Person>();
             _id = id;
+
             if (_id > 0)
-            {
-                MapPerson();
-            }
-            Email = new ValidatableObject<string>();
+                _ = MapPerson();
+
             AddValidations();
+
             SaveCommand = new Command(async () => await Save());
             DeleteCommand = new Command(async () => await Delete());
         }
 
-        private int _id;
+        private readonly int _id;
         private Person _person;
         public ICommand SaveCommand { private set; get; }
         public ICommand DeleteCommand { private set; get; }
 
-        public ValidatableObject<string> Email { get; set; } = new ValidatableObject<string>();
+        public bool Editable => _id > 0;
+        public string Title => Editable ? "Editar" : "Adicionar";
+
         public ValidatableObject<string> Name { get; set; } = new ValidatableObject<string>();
         public ValidatableObject<string> PhoneNumber { get; set; } = new ValidatableObject<string>();
+        public ValidatableObject<string> Email { get; set; } = new ValidatableObject<string>();
         public ValidatableObject<string> Password { get; set; } = new ValidatableObject<string>();
-        public ValidatableObject<DateTime> DtBirth { get; set; }
+        public ValidatableObject<DateTime> BirthDate { get; set; }
             = new ValidatableObject<DateTime>() { Value = DateTime.Today };
 
         private void AddValidations()
@@ -53,14 +52,14 @@ namespace XFSample.ViewModels
             Email.Validations.Add(new IsNotNullOrEmptyRule<string> { ValidationMessage = "O Email deve ser preenchido" });
             Email.Validations.Add(new IsValidEmailRule<string> { ValidationMessage = "O Email é inválido" });
 
-            DtBirth.Validations.Add(new ValidAgeRule<DateTime> { ValidationMessage = "Você deve ter 18 anos ou mais" });
+            BirthDate.Validations.Add(new ValidAgeRule<DateTime> { ValidationMessage = "Você deve ter 18 anos ou mais" });
 
             PhoneNumber.Validations.Add(new IsNotNullOrEmptyRule<string> { ValidationMessage = "O Telefone deve ser preenchido" });
             PhoneNumber.Validations.Add(new MinAndMaxValidRule<string>
             {
-                ValidationMessage = "O número do telefone precisar ter entre 9 e 11 digitos",
+                ValidationMessage = "O número do telefone precisar ter entre 9 e 13 digitos",
                 MinimumLenght = 9,
-                MaximumLenght = 11
+                MaximumLenght = 13
             });
         }
 
@@ -71,7 +70,8 @@ namespace XFSample.ViewModels
             Name.Value = _person.Name;
             Email.Value = _person.Email;
             PhoneNumber.Value = _person.Phone;
-            DtBirth.Value = _person.DtBirth;
+            BirthDate.Value = _person.BirthDate;
+            Password.Value = _person.Password;
         }
 
         private async Task Save()
@@ -83,7 +83,8 @@ namespace XFSample.ViewModels
                 if (!CanSave())
                     return;
 
-                _person = new Person(Name.Value, PhoneNumber.Value, Email.Value, Password.Value, DtBirth.Value);
+                _person = new Person(Name.Value, PhoneNumber.Value, Email.Value, Password.Value, BirthDate.Value, _id);
+                
                 await _repository.SaveAsync(_person);
                 await Navigation.PopAsync();
             }
@@ -123,6 +124,9 @@ namespace XFSample.ViewModels
         {
             try
             {
+                var confirm = await DialogsHelper.ShowConfirm("Atenção", "Deseja excluir este registro?");
+                if (!confirm) return;
+
                 await _repository.DeleteAsync(_person);
                 await Navigation.PopAsync();
             }
